Added skipping nodes that are in closed set in a_star algorithm.
