+++ classes/node.py
@@ -8,6 +8,7 @@ from classes.edge import Edge
 class Node:
     def __init__(self, name: str, x: float, y: float, neighbours: list[Edge] = None,
                  g: float = math.inf, f: float = math.inf, previous: "Node" = None):
+        # Name should be unique
         self.name: str = name
         self.x: float = x
         self.y: float = y
